Prototype user input added
Make a way for interacting with a "cue ball". the GameInputInit function takes reference to the cue ball object, if moving the code, init function needs to be run and keyboard functions need to be assigned before glut loop is entered.

Moved gui init func into main init function as well as the GameInputInit function.

Removed other keyboard function.